GPU-1618: Move keycloak service in a shared module, remove code dublications, add i18n
