+++ cms-ui/apps/admin-ui/src/app/app.module.ts
@@ -3,9 +3,9 @@ import { DashboardModule } from '@admin-ui/dashboard/dashboard.module';
 import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { AngularSvgIconModule } from 'angular-svg-icon';
+import { KeycloakService } from '../../../../libs/cms-components/src/lib/core/providers/keycloak/keycloak.service';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { KeycloakService } from './login/providers/keycloak/keycloak.service';
 
 const PROVIDERS: any[] = [
     { provide: LocationStrategy, useClass: HashLocationStrategy },
