+++ cms-ui/apps/admin-ui/src/app/login/components/login/login.component.html
@@ -8,7 +8,7 @@
                 <gtx-button size="large" submit [disabled]="!loginForm.valid" type="submit">{{ 'dashboard.log_in' | i18n }}</gtx-button>
             </form>
         </div>
-        <gtx-single-sign-on></gtx-single-sign-on>
+        <gtx-single-sign-on/>
     </div>
 
     <div *ngIf="keycloakError" class="keycloak-error">
