+++ cms-ui/apps/admin-ui/src/app/login/components/login/login.component.scss
@@ -37,7 +37,6 @@
     }
 }
 
-
 .keycloak-error {
     border-left: 5px solid $gtx-color-warning;     
     margin-top: 1rem;
