+++ cms-ui/apps/admin-ui/src/app/login/login.module.ts
@@ -1,11 +1,10 @@
-import {NgModule} from '@angular/core';
-import {RouterModule, provideRouter, withComponentInputBinding} from '@angular/router';
-
-import {SharedModule} from '../shared/shared.module';
-import {LoginComponent} from './components/login/login.component';
-import {SingleSignOnComponent} from './components/single-sign-on/single-sign-on.component';
-import {loginRoutes} from './login.routes';
-import {KeycloakService} from './providers/keycloak/keycloak.service';
+import { NgModule } from '@angular/core';
+import { RouterModule, provideRouter, withComponentInputBinding } from '@angular/router';
+import { KeycloakService } from '../../../../../libs/cms-components/src/lib/core/providers/keycloak/keycloak.service';
+import { SharedModule } from '../shared/shared.module';
+import { LoginComponent } from './components/login/login.component';
+import { SingleSignOnComponent } from './components/single-sign-on/single-sign-on.component';
+import { loginRoutes } from './login.routes';
 
 const COMPONENTS = [
     LoginComponent,
