+++ cms-ui/apps/editor-ui/src/app/core/core.module.ts
@@ -2,6 +2,7 @@ import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 import { APP_INITIALIZER, ErrorHandler as NgErrorHandler, NgModule, Optional, SkipSelf } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { Router } from '@angular/router';
 import { CmsComponentsModule, GCMS_COMMON_LANGUAGE, GCMS_UI_SERVICES_PROVIDER } from '@gentics/cms-components';
 import { GcmsUiLanguage } from '@gentics/cms-integration-api-models';
 import { GCMSRestClientModule, GCMSRestClientService } from '@gentics/cms-rest-client-angular';
@@ -9,6 +10,7 @@ import { GCMS_API_BASE_URL, GCMS_API_ERROR_HANDLER, GCMS_API_SID, GcmsRestClient
 import { DateTimePickerFormatProvider, GenticsUICoreModule } from '@gentics/ui-core';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { HotkeyModule } from 'angular2-hotkeys';
+import { KeycloakService } from 'libs/cms-components/src/lib/core/providers/keycloak/keycloak.service';
 import { Observable } from 'rxjs';
 import { environment as environmentConfig } from '../../environments/environment';
 import { API_BASE_URL } from '../common/utils/base-urls';
@@ -22,7 +24,6 @@ import { EmbeddedToolsService } from '../embedded-tools/providers/embedded-tools
 import { ExposedUIAPI } from '../embedded-tools/providers/exposed-ui-api/exposed-ui-api.service';
 import { ToolApiChannelService } from '../embedded-tools/providers/tool-api-channel/tool-api-channel.service';
 import { ToolMessagingChannelFactory } from '../embedded-tools/providers/tool-messaging-channel/tool-messaging-channel.factory';
-import { KeycloakService } from '../login/providers/keycloak/keycloak.service';
 import { SharedModule } from '../shared/shared.module';
 import { ApplicationStateService } from '../state';
 import { StateModule } from '../state/state.module';
@@ -85,7 +86,7 @@ export const getSidFromAppState = (appState: ApplicationStateService): Observabl
 export const createLanguageObservable = (appState: ApplicationStateService): Observable<GcmsUiLanguage> =>
     appState.select(state => state.ui.language);
 
-export function initializeApp(appState: ApplicationStateService, client: GCMSRestClientService): () => Promise<any> {
+export function initializeApp(appState: ApplicationStateService, client: GCMSRestClientService, router: Router): () => Promise<any> {
     return () => {
         // eslint-disable-next-line @typescript-eslint/no-unsafe-call
         client.init({
@@ -98,7 +99,7 @@ export function initializeApp(appState: ApplicationStateService, client: GCMSRes
             // eslint-disable-next-line @typescript-eslint/no-unsafe-call
             client.setSessionId(sid);
         });
-        return KeycloakService.checkKeycloakAuth();
+        return KeycloakService.checkKeycloakAuth(router);
     };
 }
 
@@ -183,7 +184,7 @@ const PROVIDERS = [
     {
         provide: APP_INITIALIZER,
         useFactory: initializeApp,
-        deps: [ ApplicationStateService, GCMSRestClientService ],
+        deps: [ ApplicationStateService, GCMSRestClientService, Router ],
         multi: true,
     },
 ];
