+++ cms-ui/apps/editor-ui/src/app/core/providers/i18n/translations/shared.translations.yml
@@ -1,0 +1,4 @@
+
+keycloak_not_available:
+  en: Something went wrong while accessing Keycloak. Using login without Single Sign-On instead.
+  de: Beim Zugriff auf Keycloak ist etwas schief gelaufen. Es wird stattdessen die Anmeldung ohne Single Sign-On verwendet.
