+++ cms-ui/apps/editor-ui/src/app/login/components/login/login.tpl.html
@@ -8,6 +8,10 @@
                 <gtx-button size="large" submit [disabled]="!loginForm.valid" type="submit">{{ 'user.log_in' | i18n }}</gtx-button>
             </form>
         </div>
-        <single-sign-on></single-sign-on>
+        <single-sign-on/>
+    </div>
+    
+    <div *ngIf="keycloakError" class="keycloak-error">
+        {{ keycloakError | i18n }}
     </div>
 </div>
