+++ cms-ui/apps/editor-ui/src/app/login/login.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { KeycloakService } from '../../../../../libs/cms-components/src/lib/core/providers/keycloak/keycloak.service';
 import { SharedModule } from '../shared/shared.module';
 import { LoginComponent } from './components/login/login.component';
 import { SingleSignOn } from './components/single-sign-on/single-sign-on.component';
 import { LOGIN_ROUTES } from './login.routes';
-import { KeycloakService } from './providers/keycloak/keycloak.service';
 
 const COMPONENTS = [
     LoginComponent,
